+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## v0.0.14 - In progress
 ### Added
 - Redirect to the correct entity type, if we've landed on the wrong one.
+- Add static/public to dockerignore
+- Add stage to dockerfile to build static
+### Changed
+- Moved files needed to build static inside context
 
 ## [v0.0.13](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.13) - 2020/05/03
 ### Added