+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Redirect to the correct entity type, if we've landed on the wrong one.
 
+### Changed
+- Upgrade prov-vis.
+
 ## [v0.0.13](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.13) - 2020/05/03
 ### Added
 - Added favicon